Moved & renamed users.ts to /utils folder
