+++ src/utils/authService.ts
@@ -1,26 +1,26 @@
 import { BaseDirectory, readTextFile } from "@tauri-apps/api/fs"
-import useNotification from "../../hooks/useNotification"
-import { UserType } from "../../types"
-import defaultUsersData from "../defaults/defaultUsers.json"
-import { createFolder, writeInto } from "../../utils/fileManager"
-
+import useNotification from "../hooks/useNotification"
+import { UserType } from "../types"
+import defaultUsersData from "../data/defaults/defaultUsers.json"
+import { createFolder, writeInto } from "./fileManager"
 
 /**
  * Save all users and current user data that is logged in to the app into
  * 'auth/users.data' file
- * 
+ *
  * @param currentUser The current user that is logged in to the app
  * @param allUsers List of all users to be saved to 'users.data' file
  */
-export const saveUsersData = async (currentUser: UserType, allUsers: UserType[] | any) => {
-
+export const saveUsersData = async (
+  currentUser: UserType,
+  allUsers: UserType[] | any
+) => {
   // Create 'data' folder (if it doesn't exist yet)
   await createFolder("auth")
 
   // Write into 'users.data' file
   await writeInto({ currentUser: currentUser, users: allUsers }, "auth/users")
 }
-
 
 /**
  * Initiate users data by creating the 'data' folder
@@ -29,18 +29,17 @@ export const assignDefaultUsersData = async () => {
   await saveUsersData(null, defaultUsersData)
 }
 
-
 /**
  * Get all users from 'users.data' file
- * 
+ *
  * @returns List of all registered users from 'users.data' file
  */
 export const getAllUsers = async (): Promise<UserType[]> => {
   return new Promise(async (resolve, reject) => {
     await readTextFile("auth/users.data", {
-      dir: BaseDirectory.AppData
+      dir: BaseDirectory.AppData,
     })
-      .then(res => {
+      .then((res) => {
         const data = JSON.parse(res)
         resolve(data.users)
       })
@@ -52,32 +51,30 @@ export const getAllUsers = async (): Promise<UserType[]> => {
   })
 }
 
-
 /**
  * Get a user data by its id
- * 
+ *
  * @param userId The id of the user to be searched
  * @returns The user with the specified id
  */
 export const getUserById = async (userId: string): Promise<UserType> => {
   const users = await getAllUsers()
-  const user = users.find(user => user?.id === userId)
+  const user = users.find((user) => user?.id === userId)
 
   return user ? user : null
 }
 
-
 /**
  * Get the data of current user that is logged in to the app
- * 
+ *
  * @returns The current user that is logged in to the app
  */
 export const getCurrentUser = async (): Promise<UserType> => {
   return new Promise(async (resolve, reject) => {
     await readTextFile("auth/users.data", {
-      dir: BaseDirectory.AppData
+      dir: BaseDirectory.AppData,
     })
-      .then(res => {
+      .then((res) => {
         const data = JSON.parse(res)
         resolve(data.currentUser)
       })
@@ -89,10 +86,9 @@ export const getCurrentUser = async (): Promise<UserType> => {
   })
 }
 
-
 /**
  * Set the current user that is logged in to the app
- * 
+ *
  * @param user The current user that is logged in to the app
  */
 export const setCurrentUser = async (user: UserType) => {
@@ -100,11 +96,10 @@ export const setCurrentUser = async (user: UserType) => {
   saveUsersData(user, allUsers)
 }
 
-
 /**
- * Override the list of all registered users in 'users.data' file. Doesn't 
+ * Override the list of all registered users in 'users.data' file. Doesn't
  * affect the current user
- * 
+ *
  * @param users List of all registered users to be saved to 'users.data' file
  */
 export const setAllUsers = async (users: UserType[]) => {
@@ -112,17 +107,18 @@ export const setAllUsers = async (users: UserType[]) => {
   saveUsersData(currentUser, users)
 }
 
-
 /**
  * Register a new user and save it to 'users.data' file
- * 
+ *
  * @param newUser The new user to be registered
  */
 export const registerNewUser = async (newUser: UserType) => {
-
   // Check for existing user with the same id
   if (await getUserById(newUser!.id)) {
-    useNotification("Pengguna dengan ID tersebut sudah terdaftar", "Silahkan gunakan ID lain!")
+    useNotification(
+      "Pengguna dengan ID tersebut sudah terdaftar",
+      "Silahkan gunakan ID lain!"
+    )
 
     return
   }
@@ -132,15 +128,14 @@ export const registerNewUser = async (newUser: UserType) => {
   setAllUsers(allUsers)
 }
 
-
 /**
  * Update a user data
- * 
+ *
  * @param updatedUser The updated user data to be saved
  */
 export const updateUser = async (updatedUser: UserType) => {
   const allUsers = await getAllUsers()
-  const newUsers = allUsers.map(user => {
+  const newUsers = allUsers.map((user) => {
     if (user!.id === updatedUser!.id) {
       return updatedUser
     }
@@ -149,14 +144,13 @@ export const updateUser = async (updatedUser: UserType) => {
   setAllUsers(newUsers)
 }
 
-
 /**
  * Delete a user by its id
- * 
+ *
  * @param userId The id of the user to be deleted
  */
 export const deleteUser = async (userId: string) => {
   const allUsers = await getAllUsers()
-  const newUsers = allUsers.filter(user => user!.id !== userId)
+  const newUsers = allUsers.filter((user) => user!.id !== userId)
   setAllUsers(newUsers)
 }
